Move `FileScope` to own file
